+++ docs/assets/plugins/metadata/metadata.js
@@ -48,7 +48,9 @@
                 <td>
                   ${escapeHtml(prop.description)}
                 </td>
-                <td style="text-align: center;">${prop.reflects ? '<sl-icon label="yes" name="check"></sl-icon>' : ''}</td>
+                <td style="text-align: center;">${
+                  prop.reflects ? '<sl-icon label="yes" name="check-lg"></sl-icon>' : ''
+                }</td>
                 <td>${prop.type?.text ? `<code>${escapeHtml(prop.type?.text || '')}</code>` : '-'}</td>
                 <td>${prop.default ? `<code>${escapeHtml(prop.default)}</code>` : '-'}</td>
               </tr>
@@ -114,7 +116,9 @@
                   ${
                     method.parameters?.length
                       ? `
-                        <code>${escapeHtml(method.parameters.map(param => `${param.name}: ${param.type && param.type.text ? param.type.text : ''}`).join(', '))}</code>
+                        <code>${escapeHtml(
+                          method.parameters.map(param => `${param.name}: ${param.type?.text || ''}`).join(', ')
+                        )}</code>
                       `
                       : '-'
                   }