+++ src/components/drawer/drawer.ts
@@ -2,7 +2,6 @@ import { html, LitElement } from 'lit';
 import { customElement, property, query } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
-import '../../components/icon-button/icon-button';
 import { animateTo, stopAnimations } from '../../internal/animate';
 import { emit, waitForEvent } from '../../internal/event';
 import Modal from '../../internal/modal';
@@ -12,6 +11,7 @@ import { uppercaseFirstLetter } from '../../internal/string';
 import { watch } from '../../internal/watch';
 import { getAnimation, setDefaultAnimation } from '../../utilities/animation-registry';
 import { LocalizeController } from '../../utilities/localize';
+import '../icon-button/icon-button';
 import styles from './drawer.styles';
 import type { CSSResultGroup } from 'lit';
 
@@ -22,7 +22,7 @@ import type { CSSResultGroup } from 'lit';
  * @dependency sl-icon-button
  *
  * @slot - The drawer's content.
- * @slot label - The drawer's label. Alternatively, you can use the label prop.
+ * @slot label - The drawer's label. Alternatively, you can use the `label` attribute.
  * @slot footer - The drawer's footer, usually one or more buttons representing various options.
  *
  * @event sl-show - Emitted when the drawer opens.
@@ -82,7 +82,8 @@ export default class SlDrawer extends LitElement {
 
   /**
    * The drawer's label as displayed in the header. You should always include a relevant label even when using
-   * `no-header`, as it is required for proper accessibility.
+   * `no-header`, as it is required for proper accessibility. If you need to display HTML, you can use the `label` slot
+   * instead.
    */
   @property({ reflect: true }) label = '';