+++ src/components/icon/library.ts
@@ -1,8 +1,8 @@
 import defaultLibrary from './library.default';
 import systemLibrary from './library.system';
-import type SlIcon from '../../components/icon/icon';
-import bootstrapLibrary from './library.bootstrap';
 
+import type SlIcon from '../icon/icon';
+import bootstrapLibrary from './library.bootstrap';
 
 export type IconLibraryResolver = (name: string) => string;
 export type IconLibraryMutator = (svg: SVGElement) => void;