+++ src/components/menu/menu.ts
@@ -3,7 +3,7 @@ import { customElement, query, property } from 'lit/decorators.js';
 import { emit } from '../../internal/event';
 import { getTextContent } from '../../internal/slot';
 import styles from './menu.styles';
-import type SlMenuItem from '../../components/menu-item/menu-item';
+import type SlMenuItem from '../menu-item/menu-item';
 import type { CSSResultGroup } from 'lit';
 export interface MenuSelectEventDetail {
   item: SlMenuItem;
@@ -37,9 +37,17 @@ export default class SlMenu extends LitElement {
     this.setAttribute('role', 'menu');
   }
 
-  getAllItems() {
+  getAllItems(options: { includeDisabled: boolean } = { includeDisabled: true }) {
     return [...this.defaultSlot.assignedElements({ flatten: true })].filter((el: HTMLElement) => {
-      return el.getAttribute('role') === 'menuitem';
+      if (el.getAttribute('role') !== 'menuitem') {
+        return false;
+      }
+
+      if (!options.includeDisabled && (el as SlMenuItem).disabled) {
+        return false;
+      }
+
+      return true;
     }) as SlMenuItem[];
   }
 
@@ -48,7 +56,7 @@ export default class SlMenu extends LitElement {
    * The menu item may or may not have focus, but for keyboard interaction purposes it's considered the "active" item.
    */
   getCurrentItem() {
-    return this.getAllItems().find(i => i.getAttribute('tabindex') === '0');
+    return this.getAllItems({ includeDisabled: false }).find(i => i.getAttribute('tabindex') === '0');
   }
 
   /**
@@ -56,8 +64,8 @@ export default class SlMenu extends LitElement {
    * `tabindex="-1"` to all other items. This method must be called prior to setting focus on a menu item.
    */
   setCurrentItem(item: SlMenuItem) {
-    const items = this.getAllItems();
-    const activeItem = item ?? items[0];
+    const items = this.getAllItems({ includeDisabled: false });
+    const activeItem = item.disabled ? items[0] : item;
 
     emit(this, 'sl-item-active', { detail: item });
 
@@ -75,7 +83,7 @@ export default class SlMenu extends LitElement {
    */
   typeToSelect(event: KeyboardEvent) {
 
-    const items = this.getAllItems();
+    const items = this.getAllItems({ includeDisabled: false });
 
     clearTimeout(this.typeToSelectTimeout);
     this.typeToSelectTimeout = window.setTimeout(() => (this.typeToSelectString = ''), 1000);
@@ -129,7 +137,7 @@ export default class SlMenu extends LitElement {
 
     // Move the selection when pressing down or up
     if (['ArrowDown', 'ArrowUp', 'Home', 'End'].includes(event.key)) {
-      const items = this.getAllItems();
+      const items = this.getAllItems({ includeDisabled: false });
       const activeItem = this.getCurrentItem();
       let index = activeItem ? items.indexOf(activeItem) : 0;
 
@@ -172,7 +180,7 @@ export default class SlMenu extends LitElement {
   }
 
   handleSlotChange() {
-    const items = this.getAllItems();
+    const items = this.getAllItems({ includeDisabled: false });
 
     // Reset the roving tab index when the slotted items change
     if (items.length > 0) {