+++ src/components/radio/radio.ts
@@ -1,11 +1,7 @@
 import { html, LitElement } from 'lit';
-import { customElement, property, query, state } from 'lit/decorators.js';
+import { customElement, property, state } from 'lit/decorators.js';
 import { classMap } from 'lit/directives/class-map.js';
-import { ifDefined } from 'lit/directives/if-defined.js';
-import { live } from 'lit/directives/live.js';
-import { defaultValue } from '../../internal/default-value';
-import { emit } from '../../internal/event';
-import { FormSubmitController } from '../../internal/form';
+import { emit } from 'src/internal/event';
 import { watch } from '../../internal/watch';
 import styles from './radio.styles';
 import type { CSSResultGroup } from 'lit';
@@ -18,7 +14,6 @@ import type { CSSResultGroup } from 'lit';
  * @slot - The radio's label.
  *
  * @event sl-blur - Emitted when the control loses focus.
- * @event sl-change - Emitted when the control's checked state changes.
  * @event sl-focus - Emitted when the control gains focus.
  *
  * @csspart base - The component's internal wrapper.
@@ -30,18 +25,8 @@ import type { CSSResultGroup } from 'lit';
 export default class SlRadio extends LitElement {
   static styles: CSSResultGroup = styles;
 
-  @query('.radio__input') input: HTMLInputElement;
-
-  protected readonly formSubmitController = new FormSubmitController(this, {
-    value: (control: SlRadio) => (control.checked ? control.value || 'on' : undefined),
-    defaultValue: (control: SlRadio) => control.defaultChecked,
-    setValue: (control: SlRadio, checked: boolean) => (control.checked = checked)
-  });
-
+  @state() checked = false;
   @state() protected hasFocus = false;
-
-  /** The radio's name attribute. */
-  @property() name: string;
 
   /** The radio's value attribute. */
   @property() value: string;
@@ -49,89 +34,54 @@ export default class SlRadio extends LitElement {
   /** Disables the radio. */
   @property({ type: Boolean, reflect: true }) disabled = false;
 
-  /** Draws the radio in a checked state. */
-  @property({ type: Boolean, reflect: true }) checked = false;
-
-  /**
-   * This will be true when the control is in an invalid state. Validity in radios is determined by the message provided
-   * by the `setCustomValidity` method.
-   */
-  @property({ type: Boolean, reflect: true }) invalid = false;
-
-  /** Gets or sets the default value used to reset this element. The initial value corresponds to the one originally specified in the HTML that created this element. */
-  @defaultValue('checked')
-  defaultChecked = false;
-
   connectedCallback(): void {
     super.connectedCallback();
-    this.setAttribute('role', 'radio');
-  }
-
-  /** Simulates a click on the radio. */
-  click() {
-    this.input.click();
-  }
-
-  /** Sets focus on the radio. */
-  focus(options?: FocusOptions) {
-    this.input.focus(options);
-  }
-
-  /** Removes focus from the radio. */
-  blur() {
-    this.input.blur();
-  }
-
-  /** Checks for validity and shows the browser's validation message if the control is invalid. */
-  reportValidity() {
-    return this.input.reportValidity();
-  }
-
-  /** Sets a custom validation message. If `message` is not empty, the field will be considered invalid. */
-  setCustomValidity(message: string) {
-    this.input.setCustomValidity(message);
-    this.invalid = !this.input.checkValidity();
-  }
-
-  handleBlur() {
-    this.hasFocus = false;
-    emit(this, 'sl-blur');
-  }
-
-  handleClick() {
-    if (!this.disabled) {
-      this.checked = true;
-    }
-  }
-
-  handleFocus() {
-    this.hasFocus = true;
-    emit(this, 'sl-focus');
+    this.setInitialAttributes();
+    this.addEventListeners();
   }
 
   @watch('checked')
   handleCheckedChange() {
     this.setAttribute('aria-checked', this.checked ? 'true' : 'false');
-
-    if (this.hasUpdated) {
-      emit(this, 'sl-change');
-    }
+    this.setAttribute('tabindex', this.checked ? '0' : '-1');
   }
 
   @watch('disabled', { waitUntilFirstUpdate: true })
   handleDisabledChange() {
     this.setAttribute('aria-disabled', this.disabled ? 'true' : 'false');
+  }
 
-    // Disabled form controls are always valid, so we need to recheck validity when the state changes
-    if (this.hasUpdated) {
-      this.input.disabled = this.disabled;
-      this.invalid = !this.input.checkValidity();
+  private handleBlur() {
+    this.hasFocus = false;
+    emit(this, 'sl-blur');
+  }
+
+  private handleClick() {
+    if (!this.disabled) {
+      this.checked = true;
     }
+  }
+
+  private handleFocus() {
+    this.hasFocus = true;
+    emit(this, 'sl-focus');
+  }
+
+  private addEventListeners() {
+    this.addEventListener('blur', () => this.handleBlur());
+    this.addEventListener('click', () => this.handleClick());
+    this.addEventListener('focus', () => this.handleFocus());
+  }
+
+  private setInitialAttributes() {
+    this.setAttribute('role', 'radio');
+    this.setAttribute('tabindex', '-1');
+    this.setAttribute('aria-disabled', this.disabled ? 'true' : 'false');
   }
 
   render() {
     return html`
-      <label
+      <span
         part="base"
         class=${classMap({
           radio: true,
@@ -140,17 +90,6 @@ export default class SlRadio extends LitElement {
           'radio--focused': this.hasFocus
         })}
       >
-        <input
-          class="radio__input"
-          type="radio"
-          name=${ifDefined(this.name)}
-          value=${ifDefined(this.value)}
-          .checked=${live(this.checked)}
-          .disabled=${this.disabled}
-          @click=${this.handleClick}
-          @blur=${this.handleBlur}
-          @focus=${this.handleFocus}
-        />
         <span part="control" class="radio__control">
           <span part="checked-icon" class="radio__icon">
             <svg viewBox="0 0 16 16">
@@ -164,7 +103,7 @@ export default class SlRadio extends LitElement {
         <span part="label" class="radio__label">
           <slot></slot>
         </span>
-      </label>
+      </span>
     `;
   }
 }